Extract visible multi-agent launcher seam
